Added files via upload
This is a way to make the output of the parser in order to post order, but this is just the current work, it need to be fit in the parser and there is a problem that some symbols like /d /t need to be fix, also this is just the demo, so there would be some bugs need to be fix when it can be applied to the parser's post order output. We need to discuss that.